+++ pyticketswitch/mixins.py
@@ -111,8 +111,10 @@ class PaginationMixin(object):
     def from_api_data(cls, data, *args, **kwargs):
         inst = super(PaginationMixin, cls).from_api_data(data)
 
-        results = data.get('results', {})
-        paging_data = results.get('paging_status', {})
+        paging_data = data.get('paging_status')
+        if not paging_data:
+            results = data.get('results', {})
+            paging_data = results.get('paging_status', {})
 
         inst.page_length = paging_data.get('page_length')
         inst.page_number = paging_data.get('page_number')