+++ tests/test_mixins.py
@@ -177,6 +177,45 @@ class TestPaginationMixin:
 
         assert meta.is_paginated() is False
 
+    def test_from_api_data_when_not_inside_results(self):
+        """Test from_api_data works when the data is not in `results`
+
+        Some calls (e.g. related_events.v1) have multiple pagination data
+        sections inside dicts that are not keyed as 'results'. In this case,
+        we can pass the subdict they are in directly, so we should make sure
+        the method can find the data in the base dictionary.
+        """
+
+        # state
+
+        data = {
+            'paging_status': {
+                'page_length': 50,
+                'page_number': 2,
+                'pages_remaining': 3,
+                'results_remaining': 150,
+                'total_unpaged_results': 250,
+            }
+        }
+
+        class FakeBaseMeta(object):
+
+            @classmethod
+            def from_api_data(cls, data):
+                return cls()
+
+        class FakeMeta(PaginationMixin, FakeBaseMeta):
+            pass
+
+        meta = FakeMeta.from_api_data(data)
+
+        assert meta.page_length == 50
+        assert meta.page_number == 2
+        assert meta.pages_remaining == 3
+        assert meta.results_remaining == 150
+        assert meta.total_results == 250
+
+
 
 class TestSeatPricingMixin: